install: Factor out configuration into new file
Our install config is going to grow; the "nested module in existing file"
only scales to a certain point, and we are going to pass it.
